Fix the floating-point holes: NaN ordering, float compound assignment, long double parameters, unsigned conversions, and printf's %f/%e/%g.
Compiling a printf that formats floats exercised the float paths far harder
than the suite did, and each one it touched was broken:

- ucomisd reports "unordered" as ZF=PF=CF=1, so every NaN compared equal and
  a<b/a<=b (setb/setbe) came out true.  Equality now folds in the parity flag
  and the ordered relations are evaluated as b>a / b>=a.
- x += y on a float lowered to the integer IR_ADD (and x++ added the integer
  1), operating on the bit pattern; *= and /= were rejected outright.
- long double parameters were left unpinned, so mem2reg promoted them into the
  GP file and reloaded them with fild.
- fild/cvtsi2sd are signed both ways: 64-bit unsigned conversions needed the
  2^63 split in each direction, and int->long double always used fild dword.
- A float-typed global folded its initializer as an integer, storing 1 rather
  than 1.0, and rejected 2.5 outright.
- -x on a float was 0.0 - x, which turns -0.0 into +0.0.
- Float varargs past XMM0-7 were fetched from the GP file, so the 9th float
  onwards arrived as garbage.

rt/printf.c grows %f/%e/%g plus the flags (-+ 0#), * width/precision and o/X
conversions that now run through one shared padding path, and rt/stdlib.c
gains strtoul/strtoull (the parser needs strtoull for integer suffixes).

The expansion deliberately stops at 18 significant digits, which covers a
double; README records that %f on 1e300 therefore ends in zeros rather than
glibc's exact binary value.  Only powers of ten that a double holds exactly
appear as literals in rt/, so the bootstrap stays a fixed point no matter
which compiler's strtold rounded them.

+++ include/fakecc/ir.h
@@ -84,7 +84,7 @@ typedef struct {
     IROpcode op;
     IRValue  dst;      /* meaningless for IR_RETURN/IR_BR/IR_LABEL, fill -1 */
     IRValue  a, b;     /* source operands; IR_CONST only uses imm */
-    int      imm;      /* CONST value, LABEL id, BR/CBR target, PARAM index */
+    int64_t  imm;      /* CONST value, LABEL id, BR/CBR target, PARAM index */
     SourceLoc loc;
     /* IR_CALL only: callee name + argument SSA values.  NULL for other ops. */
     char    *call_name;
